AI_Projects/Multimodal_Pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asn1crypto==1.5.1
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.76
botocore==1.40.76
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
datasets==4.4.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
email-validator==2.3.0
faiss-cpu==1.13.0
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.72.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.36.0
identify==2.6.15
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.3.0
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.8
langchain-aws==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.7
langchain-experimental==0.4.0
langchain-mcp-adapters==0.1.14
langchain-ollama==1.0.0
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langfuse==3.10.5
langgraph==1.0.3
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.33
logging==0.4.9.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.15.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
nodeenv==1.9.1
npx==0.1.6
numpy==2.3.3
ollama==0.6.1
openai==2.2.0
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pipreqs==0.4.13
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.4.0
propcache==0.4.0
protobuf==6.33.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.6
pyOpenSSL==25.3.0
pypdf==6.3.0
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
ragas==0.3.9
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
rpds-py==0.27.1
s3transfer==0.14.0
safetensors==0.7.0
scikit-learn==1.7.2
scikit-network==0.33.5
scipy==1.16.3
sentence-transformers==5.1.2
sentry-sdk==2.45.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake==1.9.0
snowflake-connector-python==4.1.1
snowflake._legacy==1.0.2
snowflake.core==1.9.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
ty==0.0.1a27
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.15
uvicorn==0.37.0
virtualenv==20.35.4
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.3
xxhash==3.6.0
yarg==0.1.10
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
